--- src/PowerShellEditorServices.Engine/Interface/ILanguageServer.cs
@@ -1,11 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Microsoft.PowerShell.EditorServices.Engine
-{
-    public interface ILanguageServer
-    {
-        Task StartAsync();
-
-        Task WaitForShutdown();
-    }
-}
