--- src/PowerShellEditorServices.Engine/Interface/ILanguageServerBuilder.cs
@@ -1,17 +1,0 @@
-using Microsoft.Extensions.Logging;
-
-namespace Microsoft.PowerShell.EditorServices.Engine
-{
-    public interface ILanguageServerBuilder
-    {
-        string NamedPipeName { get; set; }
-
-        string OutNamedPipeName { get; set; }
-
-        ILoggerFactory LoggerFactory { get; set; }
-
-        LogLevel MinimumLogLevel { get; set; }
-
-        ILanguageServer BuildLanguageServer();
-    }
-}
