--- src/PowerShellEditorServices.Engine/LanguageServer/OmnisharpLanguageServerBuilder.cs
@@ -1,40 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-
-namespace Microsoft.PowerShell.EditorServices.Engine
-{
-    public class OmnisharpLanguageServerBuilder : ILanguageServerBuilder
-    {
-        public OmnisharpLanguageServerBuilder(IServiceCollection serviceCollection)
-        {
-            Services = serviceCollection;
-        }
-
-        public bool Stdio { get; set; }
-
-        public string NamedPipeName { get; set; }
-
-        public string OutNamedPipeName { get; set; }
-
-        public ILoggerFactory LoggerFactory { get; set; } = new LoggerFactory();
-
-        public LogLevel MinimumLogLevel { get; set; } = LogLevel.Trace;
-
-        public IServiceCollection Services { get; }
-
-        public ILanguageServer BuildLanguageServer()
-        {
-            var config = new OmnisharpLanguageServer.Configuration()
-            {
-                LoggerFactory = LoggerFactory,
-                MinimumLogLevel = MinimumLogLevel,
-                NamedPipeName = NamedPipeName,
-                OutNamedPipeName = OutNamedPipeName,
-                Services = Services,
-                Stdio = Stdio
-            };
-
-            return new OmnisharpLanguageServer(config);
-        }
-    }
-}
